fix: Don't sent Authorization header if auth is None or cookies are present
This fixes problems with some SSO implementations because the
authorization header was sometimes sent (because python-requests reads
netrc if we set auth as None) or because auth is set along side
cookies.

When basic authorization header and basic auth is not permitted or the
credentials are wrong (because they are SSO credentials not for JIRA)
then the JIRA request may succeed yet the instance actually causes a
security alert raising the failedLoginCount on the JIRA instance for the
user.

Closes #37