fix: change-status getting a 404 when performing transition
There is prior logic to prevent basic auth if we have cookies (as it is
assumed we are using cookie based authentication in some SSO JIRAs).

When we request the available transitions from JIRA, we receive an XSRF
token which then results in the next request (to change status) to fail
without auth.